Move connection.{c,h}pp into libevent engine
